finished first stage of refactoring the HackYourRepo app
