+++ hackyourrepo-app/index.html
@@ -19,6 +19,6 @@
   <body>
     
 
-    <script src="./script.js"></script>
+    <script src="./util/script.js" type="module"></script>
   </body>
 </html>
